--- a/internal/wasm/sdk/internal/wasm/pool.go
+++ b/internal/wasm/sdk/internal/wasm/pool.go
@@ -172,7 +172,7 @@ func (p *Pool) SetPolicyData(ctx context.Context, policy []byte, data []byte) er
 
 		if err == nil {
 			parsedDataAddr, parsedData := vm.cloneDataSegment()
-			p.memoryMinPages = Pages(uint32(vm.memory.DataSize()))
+			p.memoryMinPages = Pages(uint32(vm.memory.DataSize(vm.store)))
 			p.vms = append(p.vms, vm)
 			p.acquired = append(p.acquired, false)
 			p.initialized = true
@@ -280,7 +280,7 @@ func (p *Pool) updateVMs(update func(vm *VM, opts vmOpts) error) error {
 				activated = true
 				policy = vm.policy
 				parsedDataAddr, parsedData = vm.cloneDataSegment()
-				seedMemorySize = Pages(uint32(vm.memory.DataSize()))
+				seedMemorySize = Pages(uint32(vm.memory.DataSize(vm.store)))
 				p.activate(policy, parsedData, parsedDataAddr, seedMemorySize)
 			}