This version of wasmtime-go reflects new Wasmtime C API specified by [Wasmtime RFC 11], and requires some changes in the way it is called.

Also, the imports that were present only to include the C headers and static lib in "`go vendor`ed" dependencies are no longer required, because these paths are now imported by wasmtime-go itself in [includebuild.go].

[Wasmtime RFC 11]: https://github.com/bytecodealliance/rfcs/pull/11
[includebuild.go]: https://github.com/bytecodealliance/wasmtime-go/blob/2f1b8a9d0/includebuild.go#L13-L17