@srenatus 2021-06-21
@olivierlemasle thanks a lot for doing this, I really appreciate your help here 😃 

The code changes look good to me, I'll play with it some locally and merge when nothing surprising happens 🚀