@srenatus 2021-06-21
What a pleasant surprise in benchmarks:

```
name                                    old time/op  new time/op  delta
WASMArrayIteration/10-16                 439µs ± 3%    86µs ± 6%  -80.49%  (p=0.008 n=5+5)
WASMArrayIteration/100-16                443µs ± 2%   109µs ± 4%  -75.34%  (p=0.008 n=5+5)
WASMArrayIteration/1000-16               573µs ± 2%   296µs ± 4%  -48.31%  (p=0.008 n=5+5)
WASMArrayIteration/10000-16             2.22ms ± 3%  1.95ms ± 3%  -12.22%  (p=0.008 n=5+5)
WASMSetIteration/10-16                   429µs ± 2%    90µs ± 4%  -79.09%  (p=0.008 n=5+5)
WASMSetIteration/100-16                  453µs ± 3%   119µs ± 3%  -73.72%  (p=0.008 n=5+5)
WASMSetIteration/1000-16                 698µs ± 5%   397µs ± 3%  -43.13%  (p=0.008 n=5+5)
WASMSetIteration/10000-16               3.80ms ± 5%  3.44ms ± 6%   -9.43%  (p=0.016 n=5+5)
WASMObjectIteration/10-16                432µs ± 4%    92µs ± 6%  -78.75%  (p=0.008 n=5+5)
WASMObjectIteration/100-16               457µs ± 3%   128µs ± 4%  -72.05%  (p=0.008 n=5+5)
WASMObjectIteration/1000-16              712µs ± 7%   429µs ± 5%  -39.77%  (p=0.008 n=5+5)
WASMObjectIteration/10000-16            3.86ms ± 2%  3.83ms ± 4%     ~     (p=0.690 n=5+5)
WASMLargeJSON/10x10-16                   488µs ± 5%   102µs ± 4%  -79.16%  (p=0.008 n=5+5)
WASMLargeJSON/10x100-16                  524µs ± 3%   148µs ± 2%  -71.77%  (p=0.008 n=5+5)
WASMLargeJSON/10x1000-16                 909µs ± 5%   648µs ± 3%  -28.69%  (p=0.008 n=5+5)
WASMLargeJSON/10x10000-16               6.07ms ±10%  5.80ms ± 5%     ~     (p=0.222 n=5+5)
WASMLargeJSON/100x100-16                 765µs ± 4%   459µs ± 4%  -39.95%  (p=0.008 n=5+5)
WASMLargeJSON/100x1000-16               3.80ms ±10%  3.47ms ± 7%     ~     (p=0.056 n=5+5)
WASMVirtualDocs/total=1/hit=1-16         446µs ± 3%    88µs ± 6%  -80.34%  (p=0.008 n=5+5)
WASMVirtualDocs/total=10/hit=1-16        446µs ± 4%    87µs ± 4%  -80.40%  (p=0.008 n=5+5)
WASMVirtualDocs/total=100/hit=1-16       445µs ± 3%   101µs ± 3%  -77.22%  (p=0.008 n=5+5)
WASMVirtualDocs/total=1000/hit=1-16      480µs ± 4%   185µs ± 4%  -61.45%  (p=0.008 n=5+5)
WASMVirtualDocs/total=10/hit=10-16       431µs ± 4%    93µs ± 3%  -78.43%  (p=0.008 n=5+5)
WASMVirtualDocs/total=100/hit=10-16      434µs ± 2%   103µs ± 8%  -76.23%  (p=0.008 n=5+5)
WASMVirtualDocs/total=1000/hit=10-16     473µs ± 3%   184µs ± 5%  -61.05%  (p=0.008 n=5+5)
WASMVirtualDocs/total=100/hit=100-16     449µs ± 4%   138µs ± 4%  -69.29%  (p=0.008 n=5+5)
WASMVirtualDocs/total=1000/hit=100-16    510µs ± 3%   212µs ± 3%  -58.49%  (p=0.008 n=5+5)
WASMVirtualDocs/total=1000/hit=1000-16   841µs ± 2%   588µs ± 3%  -30.06%  (p=0.008 n=5+5)
```

@olivierlemasle would you have any idea why it's gotten that much faster? I'm starting to wonder if I've mismanaged my `go test -bench` run here... (`go test --timeout=20m -run=^$ -tags opa_wasm,wasm_sdk_e2e ./internal/wasm/sdk/opa -bench BenchmarkWASM -count=5`)